requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-CKEditor==1.0.0
Werkzeug==3.1.3
Pillow==11.1.0
gunicorn==21.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
blinker==1.9.0
itsdangerous==2.2.0
click==8.1.8
wtforms==3.2.1
Flask-SQLAlchemy==3.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
